Documentation: Add Project Ideas document
We already had such a page on the wiki, but it's outdated and the wiki
is supposed to go the way of the dodo anyway.

This is a fresh start to make sure that all ideas we're coming up with
are still current and that there are mentors willing to support them.

Change-Id: Idd68f845930bd37a2293969b9a153cf584d6d15f
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://review.coreboot.org/c/30972
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>