mb/google/poppy/variants/atlas: move WiFi wake to GPP_B7
The latest rev. of the atlas board moves the WiFi wake source from
WAKE# to GPP_B7.  The original GPP_A0 in the device tree is just
wrong.  This also reconfigures DW1 to the GPP_B group so we can use
GPP_B7 as a wake source.

GPP_B7 is still configured as a no-connect in gpio.c, so this doesn't
actually enable WiFi wake.  We'll follow up with another patch to set
up GPP_B7 properly on boards that support it.

BUG=b:122327852
BRANCH=none
TEST=atlas still boots

Change-Id: I1816500dd0ab6186fd51aa6945faf73d00c152fe
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://review.coreboot.org/c/31211
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Furquan Shaikh <furquan@google.com>