Documentation: remove the upcoming events chapter from the conferences page
I think our docs inside of the codebase might not be the ideal place
to announce future events.

First, they might be scheduled so shortly before the conference that the
change, if at all done, would barely make it to the repo and the web. Also,
_if_ really maintained, it would churn the docs unnessesarily. But, I doubt
that anyone of us would want to maintain this here at all. Lastly, I think
that nobody out there would _look for_ upcoming events in coreboot's
documentation. We have bigger problems in the Documentation directory than
this :)

Change-Id: I918e17a427405a05722c6e0d61dc422f94cac809
Signed-off-by: Martin Kepplinger <martink@posteo.de>
Reviewed-on: https://review.coreboot.org/c/31266
Reviewed-by: Philipp Deppenwiese <zaolin.daisuki@gmail.com>
Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>